Updates according to grace comments First, the header is a landmark with a banner role. It is meant to hold site-wide content like a logo or nav. It should never contain page specific content like headings because it is for repeating content across a whole site.
Second is to make the card grid have its own container and
stick to class names as selectors instead of using element selectors and nth child.
Imagine this card grid as a component in a real site. If an author ever wanted to
add another bit of content to the page. The styles would break because
they're targeting direct children of the main landmark. It's also unnecessarily
specific using direct child nth-child in the declarations.